Strings in python

+++ calc.py
@@ -1,4 +1,4 @@
-num1=input("enter first number"))
-num2=input("enter second number"))
-sum_of_nums=int(input(num1+num2))
+num1=float(input("enter first number"))
+num2=float(input("enter second number"))
+sum_of_nums=(input(num1+num2))
 print(sum_of_nums)
